Fix runner and cache CI failures

+++ tools/slayer3d_runner.c
@@ -180,9 +180,10 @@ static bool runner_prepare_playable_map(runner_state *state, char *error_buffer,
     state->args.scene = NULL;
     state->args.player_start = NULL;
     SDL_LogInfo(SDL_LOG_CATEGORY_APPLICATION,
-                "SLAYER3D runner materialized map '%s' to '%s' (%zu playable brushes, %zu actors, player=%s)",
-                state->args.map_path, state->generated_map_output_dir, scene_desc.playable_brush_count,
-                scene_desc.actor_count, scene_desc.player_actor_id != NULL ? scene_desc.player_actor_id : "<none>");
+                "SLAYER3D runner materialized map '%s' to '%s' (%llu playable brushes, %llu actors, player=%s)",
+                state->args.map_path, state->generated_map_output_dir,
+                (unsigned long long)scene_desc.playable_brush_count, (unsigned long long)scene_desc.actor_count,
+                scene_desc.player_actor_id != NULL ? scene_desc.player_actor_id : "<none>");
     return true;
 }
